+++ packages/smarthr-ui/src/components/Button/AnchorButton.tsx
@@ -8,12 +8,12 @@ import {
   forwardRef,
   useMemo,
 } from 'react'
-import { tv } from 'tailwind-variants'
 
 import { OpenInNewTabIcon } from '../Icon'
 
 import { DisabledReason } from './DisabledReason'
-import { ActualButton } from './client'
+import { AnchorButtonInner } from './client'
+import { anchorClassNameGenerator } from './style'
 
 import type { BaseProps as ButtonProps } from './types'
 import type { ElementRef, ElementRefProps } from '../../types'
@@ -30,10 +30,6 @@ type ElementProps<T extends ElementType> = Omit<
   ComponentPropsWithoutRef<T>,
   keyof BaseProps<T> & ElementRefProps<T>
 >
-
-const classNameGenerator = tv({
-  base: 'smarthr-ui-AnchorButton',
-})
 
 const AnchorButton = forwardRef(
   <T extends ElementType = 'a'>(
@@ -54,30 +50,36 @@ const AnchorButton = forwardRef(
     }: PropsWithoutRef<BaseProps<T>> & ElementProps<T>,
     ref: Ref<ElementRef<T>>,
   ): ReactElement => {
-    const actualClassName = useMemo(() => classNameGenerator({ className }), [className])
+    const classNames = useMemo(() => {
+      const { wrapper, inner } = anchorClassNameGenerator()
+
+      return {
+        wrapper: wrapper({ variant, size, wide, className }),
+        inner: inner({ size }),
+      }
+    }, [variant, size, wide, className])
 
     // target="_blank" だが OpenInNewTabIcon を表示したくない場合 suffix に null を指定すれば表示しないようにしている
     const actualSuffix =
       target === '_blank' && !prefix && suffix === undefined ? <OpenInNewTabIcon /> : suffix
 
+    const Component = elementAs || 'a'
+
     const button = (
-      <ActualButton
+      <Component
         {...rest}
-        elementAs={elementAs}
-        anchorRef={ref}
+        ref={ref}
         href={href}
         target={target}
         rel={rel === undefined && target === '_blank' ? 'noopener noreferrer' : rel}
-        isAnchor
-        variant={variant}
-        size={size}
-        wide={wide}
-        className={actualClassName}
-        prefix={prefix}
-        suffix={actualSuffix}
+        className={classNames.wrapper}
       >
-        {children}
-      </ActualButton>
+        {prefix}
+        <AnchorButtonInner className={classNames.inner} prefix={prefix} suffix={actualSuffix}>
+          {children}
+        </AnchorButtonInner>
+        {actualSuffix}
+      </Component>
     )
 
     if (!href && inactiveReason) {
