+++ packages/smarthr-ui/src/components/Button/Button.tsx
@@ -1,18 +1,14 @@
 import { type ButtonHTMLAttributes, forwardRef, useId, useMemo } from 'react'
-import { tv } from 'tailwind-variants'
+
+import { Loader } from '../Loader'
 
 import { DisabledReason } from './DisabledReason'
 import { ActualButton, LoadingStatus } from './client'
+import { buttonClassNameGenerator } from './style'
 
 import type { BaseProps } from './types'
 
 type Props = BaseProps & Omit<ButtonHTMLAttributes<HTMLButtonElement>, keyof BaseProps>
-
-const classNameGenerator = tv({
-  slots: {
-    wrapper: 'smarthr-ui-Button',
-  },
-})
 
 export const Button = forwardRef<HTMLButtonElement, Props>(
   (
@@ -36,12 +32,14 @@ export const Button = forwardRef<HTMLButtonElement, Props>(
     const generatedId = useId()
     const buttonId = id || generatedId
     const classNames = useMemo(() => {
-      const { wrapper } = classNameGenerator()
+      const { wrapper, loader, inner } = buttonClassNameGenerator()
 
       return {
-        wrapper: wrapper({ className }),
+        wrapper: wrapper({ variant, size, wide, className }),
+        loader: loader({ variant }),
+        inner: inner({ size }),
       }
-    }, [className])
+    }, [variant, size, wide, className])
 
     const button = (
       <ActualButton
@@ -49,12 +47,11 @@ export const Button = forwardRef<HTMLButtonElement, Props>(
         buttonRef={ref}
         type={type}
         id={buttonId}
-        disabled={disabled}
-        $loading={loading}
-        variant={variant}
-        size={size}
-        wide={wide}
-        className={classNames.wrapper}
+        disabled={loading || disabled}
+        loader={
+          loading ? <Loader role="presentation" size="S" className={classNames.loader} /> : null
+        }
+        classNames={classNames}
         prefix={prefix}
         suffix={suffix}
       >
