+++ packages/smarthr-ui/src/components/Button/client/index.ts
@@ -1,2 +1,3 @@
 export { ActualButton } from './ActualButton'
+export { AnchorButtonInner } from './AnchorButtonInner'
 export { LoadingStatus } from './LoadingStatus'
